app/requirements.txt
fastapi
uvicorn
elasticsearch
structlog
ansible
paramiko
python-dotenv
pyotp
passlib
bcrypt
python-jose[cryptography]
pytest
ansible-runner
